Renamed player.loadShow() to player.loadEmbeddedPlayer()
The previously named function `loadShow()`didn't really load a show, it loads the embedded URL. The embedded URL can in some cases autoplay a show, which is where the confusion started, but it not always. So to make it more clear we needed to rename it to what it actually does, loading an embedded player.