Remove event model
As this project is foremost for users to quickly learn how to get up and running we decided to remove the event models. Even though it is cleaner code and more Swifty to use an enum for event-based-data, it would still require an extra step for the developer implementing this solution. Trying to keep it simple is more important in this repository.